Refactor MDXComponents: move component definitions to a new file for improved organization and maintainability
